Style fix: Fix style issues for all files
