+++ build.gradle
@@ -1,8 +1,14 @@
 plugins {
     id 'java'
     id 'application'
+    id 'checkstyle'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
 }
+
+checkstyle {
+    toolVersion = '10.2'
+}
+
 
 repositories {
     mavenCentral()
