+++ src/main/java/geng/commands/AddEventCommand.java
@@ -1,7 +1,12 @@
-package duke.commands;
+package geng.commands;
 
-import duke.tasks.*;
-import duke.ui.*;
+import geng.tasks.Events;
+import geng.tasks.Task;
+import geng.tasks.TaskList;
+import geng.ui.GengException;
+import geng.ui.Storage;
+import geng.ui.Ui;
+
 
 /**
  * Represents a command to add an event task.
