+++ src/main/java/geng/commands/DeleteCommand.java
@@ -1,7 +1,11 @@
-package duke.commands;
+package geng.commands;
 
-import duke.tasks.*;
-import duke.ui.*;
+import geng.tasks.Task;
+import geng.tasks.TaskList;
+import geng.ui.GengException;
+import geng.ui.Storage;
+import geng.ui.Ui;
+
 
 /**
  * Represents a command to delete a task from the task list.
