+++ src/main/java/geng/commands/ListCommand.java
@@ -1,7 +1,9 @@
-package duke.commands;
+package geng.commands;
 
-import duke.tasks.*;
-import duke.ui.*;
+import geng.tasks.TaskList;
+import geng.ui.Storage;
+import geng.ui.Ui;
+
 
 /**
  * Represents a command to list all tasks in the task list.
