+++ src/main/java/geng/tasks/TaskList.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package geng.tasks;
 
 import java.util.ArrayList;
