+++ src/main/java/geng/ui/GengException.java
@@ -1,4 +1,4 @@
-package duke.ui;
+package geng.ui;
 
 /**
  * Custom exception class to handle errors in the Duke application.
@@ -6,12 +6,10 @@ package duke.ui;
  * more specific error handling in the application.
  */
 public class GengException extends Exception {
-
     /**
      * Constructs a new GengException with the specified detail message.
      *
-     * @param message The detail message, which is saved for later retrieval by the
-     * {@link Throwable#getMessage()} method.
+     * @param message The detail message of the Exception.
      */
     public GengException(String message) {
         super(message);
