+++ src/main/java/geng/ui/Parser.java
@@ -1,7 +1,15 @@
-package duke.ui;
+package geng.ui;
 
-import duke.commands.*;
-import duke.ui.GengException;
+import geng.commands.AddDeadlineCommand;
+import geng.commands.AddEventCommand;
+import geng.commands.AddTodoCommand;
+import geng.commands.Command;
+import geng.commands.DeleteCommand;
+import geng.commands.FindCommand;
+import geng.commands.ListCommand;
+import geng.commands.ListDateCommand;
+import geng.commands.MarkCommand;
+import geng.commands.UnmarkCommand;
 
 /**
  * The Parser class is responsible for parsing user input commands
