+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hey yo! I'm duke.Geng
+Hey yo! I'm geng.Geng
 How can I help you?
 Got it. I've added this task:
   [T][ ] borrow book
